pyproject.toml
[project]
name = "dify-plugin-broker"
version = "0.0.5"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "dify-plugin==0.2.1",
    "loguru==0.7.3",
    "python-dotenv==1.1.0",
    "requests==2.32.3",
    "retry2==0.9.5",
]
